Created application type icon component.
