+++ web-dashboard/src/components/common/applicationTypeIcon/index.ts
@@ -1,0 +1,3 @@
+import ApplicationTypeIcon from './ApplicationTypeIcon';
+
+export default ApplicationTypeIcon;
